deploy/docker/requirements.txt
pyyaml
requests
fastapi
uvicorn[standard]
